Time: 0 ms (100%), Space: 21.6 MB (99.76%) - LeetHub

+++ 0098-validate-binary-search-tree/0098-validate-binary-search-tree.cpp
@@ -11,7 +11,7 @@
  */
 class Solution {
 public:
-bool helper(TreeNode* root, int mini , int maxi)
+bool helper(TreeNode* root,long long mini , long long maxi)
 {
     if(root==NULL) return true;
     if( root->val >mini && root->val<maxi)
@@ -21,11 +21,11 @@ bool helper(TreeNode* root, int mini , int maxi)
         return l&&r;
     }
     return false;
-    
+
 
 }
     bool isValidBST(TreeNode* root) {
      if(root==NULL) return true;
-     return helper(root,INT_MIN ,INT_MAX);   
+     return helper(root,LONG_MIN ,LONG_MAX);   
     }
 };
